+++ core/src/main/java/Fisicas/Camara.java
@@ -6,7 +6,7 @@ import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.math.Vector3;
 import com.badlogic.gdx.utils.viewport.FitViewport;
 import com.badlogic.gdx.utils.viewport.Viewport;
-import entidades.Personaje;
+import entidades.personajes.Personaje;
 
 public class Camara {
     private OrthographicCamera camera;