+++ core/src/main/java/com/principal/AntsArmageddon.java
@@ -2,14 +2,9 @@ package com.principal;
 
 import com.badlogic.gdx.Game;
 import com.badlogic.gdx.assets.AssetManager;
-import com.badlogic.gdx.audio.Sound;
-import com.badlogic.gdx.graphics.Texture;
-import com.badlogic.gdx.graphics.g2d.TextureAtlas;
-import jdk.jshell.execution.Util;
 import managers.GestorAssets;
+import managers.ScreenManager;
 import screens.MenuScreen;
-import utils.Constantes;
-import utils.Utiles;
 
 /** {@link com.badlogic.gdx.ApplicationListener} implementation shared by all platforms. */
 public class AntsArmageddon extends Game {
@@ -20,6 +15,7 @@ public class AntsArmageddon extends Game {
     public void create() {
         GestorAssets.load();
 
+        ScreenManager.setJuego(this);
         setScreen(new MenuScreen(this));
     }