+++ core/src/main/java/screens/MenuScreen.java
@@ -2,7 +2,6 @@ package screens;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Screen;
-import com.badlogic.gdx.assets.AssetManager;
 import com.badlogic.gdx.audio.Sound;
 import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.graphics.OrthographicCamera;
@@ -11,7 +10,6 @@ import com.badlogic.gdx.graphics.g2d.Sprite;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.scenes.scene2d.ui.ImageButton;
-import com.badlogic.gdx.scenes.scene2d.ui.Skin;
 import com.badlogic.gdx.scenes.scene2d.ui.Table;
 import com.badlogic.gdx.utils.viewport.FitViewport;
 import com.principal.AntsArmageddon;
@@ -89,22 +87,10 @@ public class MenuScreen implements Screen {
         escenario.addActor(table);
     }
 
-
-
-
-    @Override
-    public void resize(int ancho, int alto) { viewport.update(ancho, alto, true); }
-
-    @Override
-    public void dispose() { escenario.dispose(); batch.dispose(); }
-
-    @Override
-    public void pause() { }
-
-    @Override
-    public void resume() { }
-
-    @Override
-    public void hide() { }
+    @Override public void resize(int ancho, int alto) { viewport.update(ancho, alto, true); }
+    @Override public void dispose() { escenario.dispose(); batch.dispose(); }
+    @Override public void pause() { }
+    @Override public void resume() { }
+    @Override public void hide() { }
 
 }